Rearranged folder structure and fixed the restaurant profile retrieving by id
