+++ src/home/HomePage.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import CardLayout from './cards/CardLayout';
-import FormModal from './forms/FormModal';
+import CardLayout from './CardLayout';
+import FormModal from './FormModal';
 
 class HomePage extends Component {
     render() {
