+++ src/login/Login.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import './Login.css';
-import { validateUser } from '../../../service/UserService';
+import { validateUser } from '../service/UserService';
 
 const useStateWithLocalStorage = localStorageKey => {
     const [value, setValue] = React.useState(
