+++ src/navbar/MyNavBar.js
@@ -1,6 +1,6 @@
-import React, { Component } from "react";
-import '../../App.css';
-import logo from '../../images/cutout_logo.png';
+import React from "react";
+import '../App.css';
+import logo from '../images/cutout_logo.png';
 import SearchForm from "./SearchForm";
 import { Link, useLocation } from 'react-router-dom';
 import { useState } from "react";
@@ -15,7 +15,7 @@ const MyNavBar = props => {
 
     return (
       <div>
-        <nav className="navbar navbar-expand-lg nav fadeIn second">
+        <nav className="navbar navbar-expand-lg nav second">
                 <Link to="/home"> <img src={logo} alt="logo" style={{ width: 65, marginTop: -7 }} className="navbar-brand title"  /></Link>
           <button className="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
             <span className="navbar-toggler-icon"></span>
