+++ src/restaurant_profile/CardLayout.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import Card from './Card';
-import { useFetch } from '../../../../service/ReviewService';
+import { useFetch } from '../../../service/ReviewService';
 
 function CardLayout(props) {
-  const { loading, data } = useFetch('getById', 1);
+    const { loading, data } = useFetch('getById', props.id);
 
   return (
       <div className="container-fluid ">
