+++ src/restaurant_profile/RestaurantImage.js
@@ -1,8 +1,9 @@
 import React from "react";
+import './container-style.css';
 const RestaurantImage = (props) => {
     return (
         <div className="overflow" >
-            <img src={props.imageURL} alt="image" className="card-img-top-column" />
+            <img src={props.imageURL} alt="image" id="profileimg" className="card-img-top-column" />
         </div>
     )
 }
